LayoutTests/fast/mediastream/RTCSessionDescription-expected.txt

Tests RTCSessionDescription.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sessionDescription = new RTCSessionDescription(); did not throw exception.
PASS sessionDescription = new RTCSessionDescription(initializer); did not throw exception.
PASS sessionDescription.type is "offer"
PASS sessionDescription.sdp is "foobar"
PASS new RTCSessionDescription(null); threw exception TypeError: Optional RTCSessionDescription constructor argument must be a valid Dictionary.
PASS new RTCSessionDescription(5); threw exception TypeError: Optional RTCSessionDescription constructor argument must be a valid Dictionary.
PASS new RTCSessionDescription('foobar'); threw exception TypeError: Optional RTCSessionDescription constructor argument must be a valid Dictionary.
PASS new RTCSessionDescription({type:'foobar', sdp:'x'}); threw exception TypeError: Invalid RTCSessionDescription constructor arguments.
PASS new RTCSessionDescription({type:'offer', sdp:''}); threw exception TypeError: Invalid RTCSessionDescription constructor arguments.
PASS new RTCSessionDescription({}); did not throw exception.
PASS new RTCSessionDescription({type:'offer', sdp:'x'}); did not throw exception.
PASS new RTCSessionDescription({type:'answer', sdp:'x'}); did not throw exception.
PASS new RTCSessionDescription({type:'pranswer', sdp:'x'}); did not throw exception.
PASS successfullyParsed is true

TEST COMPLETE
